refactor(hub-common): moves expandPredicate function and exports serialization of IQuery
affects: @esri/hub-common
