+++ packages/common/src/search/_internal/index.ts
@@ -4,3 +4,4 @@ export * from "./portalSearchItems";
 export * from "./hubSearchItems";
 export * from "./portalSearchGroups";
 export * from "./portalSearchUsers";
+export * from "./ifilter-utils";
